app/requirements.txt
Flask==2.3.2
python-docx==0.8.11
openpyxl==3.1.2